refactor(test-harness): one type per file — split toolchain naming
toolchain.rs held two enums and three free functions. Role gets its own module;
ToolchainMatrix keeps the file as the toolchain-naming type, and the free
functions become its associated functions — ToolchainMatrix::classify (already),
::pipeline_of, ::humanize_mode, and the private ::codegen_name — rather than a
Mode newtype that would own nothing.
