+++ solx-benchmark-converter/src/output/summary/summary_template.rs
@@ -18,9 +18,9 @@ use crate::output::summary::failure_verdict::FailureVerdict;
 use crate::output::summary::health_issue::HealthIssue;
 use crate::output::summary::listing_section::ListingSection;
 use crate::output::summary::output_verdict::OutputVerdict;
+use crate::output::summary::role::Role;
 use crate::output::summary::suite_row::SuiteRow;
 use crate::output::summary::suite_stats::SuiteStats;
-use crate::output::summary::toolchain::Role;
 use crate::output::summary::truncated::Truncated;
 use crate::utils::agreeing;
 use crate::utils::count_noun;
